02-monolithic-scaling/requirements.txt
flask
flask_sqlalchemy
flask_login
pymysql
werkzeug
cryptography
python-dotenv
